Edited some name files and added navigation to AddContact screen
