+++ src/Screens/AddContact.tsx
@@ -1,14 +1,7 @@
+import {View, StyleSheet} from 'react-native';
 import React from 'react';
-import Icon from 'react-native-vector-icons/Ionicons';
-import {
-  View,
-  Text,
-  StyleSheet,
-  Touchable,
-  TouchableOpacity,
-  ScrollView,
-} from 'react-native';
-import Contacto from '../Components/Contacto';
+import InputDestinatario from '../Components/InputDestinatario';
+import FormButton from '../Components/Button';
 import {useNavigation} from '@react-navigation/native';
 
 export default function AddContact() {
@@ -16,52 +9,44 @@ export default function AddContact() {
 
   return (
     <View style={styles.container}>
-      <Text style={styles.Titledestino}>Destinatario</Text>
-
-      <TouchableOpacity
-        onPress={() => {
-          navigation.navigate('AddContact');
-        }}>
-        <View style={styles.ContainerBottn}>
-          <Icon name="add-circle" size={60} color="#4A52FF" />
-          <Text style={styles.TextNuevo}>Nuevo</Text>
-        </View>
-      </TouchableOpacity>
-      <Text style={styles.Titleguardado}>Guardados</Text>
-      <ScrollView>
-        <Contacto nombre="Yahir Cortes" numero={1234} icono="cc-visa" />
-        <Contacto nombre="Roberto" numero={1234} icono="cc-mastercard" />
-        <Contacto nombre="Enrique" numero={1234} icono="cc-visa" />
-        <Contacto nombre="Ricardo Milos" numero={1234} icono="cc-mastercard" />
-        <Contacto nombre="Yahir Cortes" numero={1234} icono="cc-visa" />
-      </ScrollView>
+      <InputDestinatario
+        style={styles.Input}
+        placeholder="Nombre Completo"
+        icono="user"
+      />
+      <InputDestinatario placeholder="Alias" />
+      <InputDestinatario
+        style={styles.Inputbt}
+        placeholder="Numero de Cuenta"
+        icono="wallet"
+        tipo="number-pad"
+      />
+      <FormButton
+        text="Agregar"
+        fn={() => {
+          //add contact...
+          //pass contact to Transferir screen
+          navigation.navigate('Transferir');
+        }}
+      />
     </View>
   );
 }
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: '#fff',
+    backgroundColor: '#EEEAEA',
+    alignItems: 'center',
     padding: 20,
   },
-  Titledestino: {
-    fontSize: 20,
-    fontWeight: 'bold',
+  containerBtn: {
+    backgroundColor: '#fff',
+    marginTop: 10,
   },
-  TextNuevo: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    color: '#4A52FF',
+  Input: {
+    marginTop: 100,
   },
-  ContainerBottn: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginTop: 20,
-    gap: 20,
-  },
-  Titleguardado: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    marginTop: 20,
+  Inputbt: {
+    marginBottom: 60,
   },
 });
